+++ dialoglib/src/main/java/com/nasduck/dialoglib/interfaces/IDialogHeader.java
@@ -16,5 +16,5 @@ public interface IDialogHeader {
      * get dialog background radius
      * @param cornerRadius
      */
-    void getCornerRadius(float cornerRadius);
+    void getCornerRadius(int cornerRadius);
 }
